+++ src/utils/show-current-dir.js
@@ -1,0 +1,1 @@
+export const showCurrentDir = () => console.log(`You are currently in ${process.env.currentDir}`)
